pyloggernmea/requirements.txt
colorama==0.4.6
exceptiongroup==1.2.2
importlib-metadata==6.7.0
iniconfig==2.0.0
lockfile==0.12.2
ordered-set==4.1.0
packaging==24.0
pluggy==1.2.0
pyserial==3.5
pytest==7.4.4
python-daemon==3.1.2
tomli==2.0.1
typing-extensions==4.7.1
zipp==3.15.0
zstandard==0.21.0
